RFS now writable.
Made the root filesystem writable, which helps in order to be productive.

Kernel boots. Flash and networking seems to be working as intended.

The NAND Flash driver is now a platform driver and the partition table is
fetched from the struct declared in simcom.c.

TODO: Remove globals from simcom_nand.c