Working audio
simcom.c
--------
- Added the PCA9539 I/O expander (available through /sys/class/gpio).
- Added GPIO configuration for the I2S module (in master mode).

simcom-audio.c
--------------
The simcom ALSA ASoC machine now has its own hw_params which allows us to set the MCLK frequency of the CODEC and to configure the I2S/CODEC in master/slave mode respectively.

TODO: Some noise seems to be added to the capture signal. This should be located.